requirements.txt
psutil==7.0.0
python-dotenv==1.0.1
requests==2.32.4
